refactor: DealerResult 클래스로 리팩토링
